refactoring(15): decompose StepResources.tsx (extract stepResourcesSections)
Third slice of req-15 (multi-PR). StepResources.tsx 1558 -> 1161 lines: extract
the ORS provider model tables + inferOrsProvider (used only by the pickers) and
the OrsModelPicker + FewshotManager subcomponents VERBATIM into
components/Wizard/stepResourcesSections.tsx; container imports the two comps.
Moved code byte-identical (modulo export kw); added the new file's useCallback
import and dropped the container's now-dead React default import.

Checks: typecheck 0 non-test errors; build ok; vitest 4 passed / 2 failed
(pre-existing brittle PaperBench tests). req-15 file RETAINED (multi-PR §9/§10).
Remaining: SettingsPage + DetailPanel (monoliths — higher-risk prop-contract
refactor, not a verbatim move).
