+++ ari-core/ari/viz/frontend/src/components/README.md
@@ -77,6 +77,7 @@ Feature-grouped React components — one subdirectory per dashboard page plus sh
   - `StepGoal.tsx` — research goal / chat step.
   - `StepLaunch.tsx` — final review + launch step.
   - `StepResources.tsx` — provider/model and container-image step.
+  - `stepResourcesSections.tsx` — ORS model tables + OrsModelPicker/FewshotManager (extracted from StepResources in req 15).
   - `StepScope.tsx` — search scope (max depth/nodes) step.
   - `WizardPage.tsx` — wizard container/step orchestration.
 - `Workflow/` — workflow stages/pipeline page.
