+++ ari-core/ari/viz/frontend/src/components/Wizard/README.md
@@ -9,5 +9,6 @@ Run-launch wizard — multi-step form to configure and start a new run.
 - `StepGoal.tsx` — research goal / chat step.
 - `StepLaunch.tsx` — final review + launch step.
 - `StepResources.tsx` — provider/model and container-image step.
+- `stepResourcesSections.tsx` — ORS model tables + OrsModelPicker/FewshotManager (extracted from StepResources in req 15).
 - `StepScope.tsx` — search scope (max depth/nodes) step.
 - `WizardPage.tsx` — wizard container/step orchestration.
