+++ refactoring/COMPLETED.md
@@ -568,12 +568,19 @@ done (or the remainder is moved to a further follow-up).
   removed the now-dead reactflow Handle/Position/NodeTypes container imports.
   Checks: typecheck 0 non-test errors; build ok; vitest 4 passed / 2 failed
   (pre-existing brittle PaperBench tests). PR/commit on branch refactoring.
-- Remaining components (each a future slice): StepResources.tsx (~1558, has clean
-  module-scope units: inferOrsProvider/OrsModelPicker/FewshotManager),
-  SettingsPage.tsx (~1123) and DetailPanel.tsx (~938) — both MONOLITHS (no
-  module-scope helper/presentational units; their tab/field render blocks close
-  over heavy local state, so they need genuine prop-contract refactoring with
-  behavior verification, NOT a verbatim move — a higher-risk follow-up);
+- 2026-05-30 — **StepResources.tsx** decomposed (1558 -> 1161 lines). Extracted
+  the ORS provider model tables + inferOrsProvider helper (used only by the
+  pickers) and the OrsModelPicker + FewshotManager subcomponents VERBATIM into
+  components/Wizard/stepResourcesSections.tsx; container imports the two comps.
+  Moved code byte-identical (modulo export kw); fixed the new file's
+  useCallback import and dropped the now-dead React default import from the
+  container. Checks: typecheck 0 non-test errors; build ok; vitest 4 passed /
+  2 failed (pre-existing brittle PaperBench tests). PR/commit on branch refactoring.
+- Remaining components: SettingsPage.tsx (~1123) and DetailPanel.tsx (~938) —
+  both MONOLITHS (no module-scope helper/presentational units; their tab/field
+  render blocks close over heavy local state, so they need genuine prop-contract
+  refactoring with behavior verification, NOT a verbatim move — a higher-risk
+  follow-up);
   optional finer split of resultSections.tsx + the low/med-risk ResultsPage
   container seams (per refactoring/notes/03_resultspage_decomposition.md).
